Structuring/hierarchy improvement: split up python project into multiple modules.
