+++ helper_functions.py
@@ -1,0 +1,2 @@
+def output(x):
+    print(x)
